Add files via upload
This document outlines the payment journey that you, as a user can experience in the web space. When you have multiple interfaces from where you can subscribe to a plan, the web journey is called PWA (Progressive Web App). The experience you have on mobile web is like that on a mobile app.