Add files via upload
This user manual is meant as an introductory guide for a user to use the PerfOptima system. The main aim is to
provide a self-learning guide on how to use the features and functionality of PerfOptima. It assumes that the user has
prior knowledge of the various terminologies and controls mentioned in the system as well as a basic idea of how to
manipulate those controls.
After going through the guide, the reader should be in a position to start using PerfOptima for Portfolio
Management and Reporting. In particular, he should be able to:
• Access the system website through its url
• Manipulate and manage the various controls on the system
• Produce and print various reports in an Excel format
• Advance to become a power user by further self-learning and exploring.